src/requirements.txt
# core dependencies
python-docx>=0.8.11

# optional dependencies (local LLM inference)
llama-cpp-python>=0.2.23

# optional dependencies (remote API call)
openai>=1.12.0